+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 
 # Changelog - SchoolServices - ConstHistCert
 
+## [2.6.1] - 2025-11-07
+René Alejandro Rivas
+
+### Changed:
+- qr-validation.model:
+  - 'delivery' field added to QrValidation model
+- qr-validation.repo:
+  - 'delivery' attribute added to saveProcedureData()
+- academic-record.service:
+  - in setAcademicRecordServiceProperties() validation for delivery type removed
+- academic-record.controller:
+  - POST: validation for delivery type removed
+
+
 ## [2.6.0] - 2024-06-20
 René Alejandro Rivas